loggerMock.js creation@
