+++ src/app/containers/Header/index.test.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import { string, shape } from 'prop-types';
+import { shouldMatchSnapshot } from '@bbc/psammead-test-helpers';
 import HeaderContainer from './index';
 import { RequestContextProvider } from '#contexts/RequestContext';
 import { ServiceContext } from '#contexts/ServiceContext';
 import { ToggleContext } from '#contexts/ToggleContext';
-import { shouldMatchSnapshot } from '#testHelpers';
 import { service as pidginServiceConfig } from '#lib/config/services/pidgin';
 
 const defaultToggleState = {
