+++ src/server/utilities/logResponseTime/index.test.js
@@ -1,4 +1,4 @@
-import { loggerMock } from '#testHelpers/index';
+import loggerMock from '#testHelpers/loggerMock';
 import logResponseTime from '.';
 
 const req = {
